DO NOT MERGE: Test Jenkins CI integration by deleting code
This commit deletes all code and assets under the `Structure_of_Phenol`
game. This is done to test what happens when an external contributor
opens a pull request on the Veativetech/VeativeWebVR repo. The expected
outcome is a failed test since this commit deletes critical code that is
used in the unit tests.

The failed test results should be reported in the pull request.

Signed-off-by: Justin W. Flory <git@jwf.io>